requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pyodbc>=5.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.18.0
pydantic>=2.0.0
llama-cpp-python>=0.2.0
pyinstaller>=6.0.0
cryptography>=41.0.0
bcrypt>=4.0.0
tensorflow>=2.15.0
holidays>=0.40
huggingface_hub>=0.20.0
requests>=2.28.0
openai>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
